GameAssistant/video_extract/requirements_video.txt
yt-dlp>=2024.1.0
openai>=1.0.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
